// src/enums/colors.ts
export enum Colors {
  WHITE = '#FFFFFF',
  BLACK = '#000000',
  TEXT = '#0C0C0C',
  GRAY = '#D8DADD',
  SURFACE = '#093545',
  SURFACE_TWO = '#EAECF0',
  TEXT_SECONDARY = '#4F4F4F',
  SECONDARY_DARK = '#777777',
  BOX_SHADOW = 'rgba(0, 0, 0, 0.02)',
  TEXT_CAPTION = '#7C8684',
  TEXT_BACKGROUND = '#E2E8F5',
  TEXT_BRAND = '#7E7E7E',
  GRAY_HOVER = '#404040',
  TEXT_TERITARY = '#7C8684',
  STROKE = 'rgba(225, 225, 225, 1)',
  ACCENT = 'rgba(37, 131, 176, 1)',
  UI = 'rgba(194, 204, 202, 1)',
}